requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic-settings
alembic
pytest
httpx
pytest-asyncio
fastapi-cache2[redis]
redis
sqladmin
fastapi-users[sqlalchemy]
celery
asyncpg
passlib[bcrypt]
bcrypt
python-jose[cryptography]
Pillow
elasticsearch
itsdangerous